티셔츠 이름 변경 기능 오류 수정, 엔티티의 필요 없는 멤버 삭제

+++ src/main/java/com/gamechanger/service/clothes/ClothesServiceImpl.java
@@ -5,9 +5,7 @@ import com.gamechanger.domain.Image;
 import com.gamechanger.domain.User;
 import com.gamechanger.repository.ClothesRepository;
 import com.gamechanger.service.image.ImageService;
-import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityNotFoundException;
-import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.json.simple.parser.ParseException;
@@ -29,17 +27,16 @@ public class ClothesServiceImpl implements ClothesService {
         Clothes clothes = Clothes.builder()
                 .clothesName(clothesName)
                 .roomId(String.valueOf(UUID.randomUUID()))
-                .background("white")
-                .shape("half")
                 .imageFileList(new HashMap<>())
                 .user(user)
                 .build();
         String[] defaultAccesses = {"room:write"};
+        Clothes savedClothes = clothesRepository.save(clothes);
         String responseRoomId = liveblocksService.createRoom(clothes.getRoomId(), defaultAccesses);
         // 룸 생성은 컨트롤러에서 하던지, save를 먼저 하던지 해야 할 듯?
         // NotNull이나 컬럼 길이 제한 같은게 레포에 save할 때 검사하는 것 같은데, 룸을 먼저 생성하니까 룸은 생성되고 옷은 생성 안되게 됨.
         log.info("Clothes Name: {}, Room Id: {} created.", clothesName, responseRoomId);
-        return clothesRepository.save(clothes);
+        return savedClothes;
     }
 
     @Override
@@ -69,7 +66,11 @@ public class ClothesServiceImpl implements ClothesService {
 
     @Override
     public void deleteClothes(Long systemClothesId) {
-        Clothes clothes = clothesRepository.findBySystemClothesId(systemClothesId).get();
+        Optional<Clothes> optionalClothes = clothesRepository.findBySystemClothesId(systemClothesId);
+        if (optionalClothes.isEmpty()) {
+            return;
+        }
+        Clothes clothes = optionalClothes.get();
         for (Image image : clothes.getImageFileList().values()) {
             imageService.deleteImage(image.getFileName());
         }
@@ -84,12 +85,17 @@ public class ClothesServiceImpl implements ClothesService {
         Clothes clothes = clothesRepository.findBySystemClothesId(systemClothesId)
                 .orElseThrow(() -> new EntityNotFoundException("Clothes not found"));
         clothes.setClothesName(newClothesName);
-        return clothesRepository.save(clothes);
+        clothesRepository.save(clothes);
+        return clothesRepository.findBySystemClothesId(systemClothesId).orElse(null);
     }
 
     @Override
     public Image uploadUserImage(Long systemClothesId, MultipartFile uploadImage, String clothesName) throws IOException {
-        Clothes clothes = clothesRepository.findBySystemClothesId(systemClothesId).get();
+        Optional<Clothes> optionalClothes = clothesRepository.findBySystemClothesId(systemClothesId);
+        if (optionalClothes.isEmpty()) {
+            return null;
+        }
+        Clothes clothes = optionalClothes.get();
         Image uploadedImage = imageService.uploadImage(clothes, uploadImage.getBytes(), "front");
         clothes.addImageFile(uploadedImage);
         return uploadedImage;
@@ -97,7 +103,11 @@ public class ClothesServiceImpl implements ClothesService {
 
     @Override
     public Image createAiImageByPrompt(Long systemClothesId, String clothesName, String style, String prompt) {
-        Clothes clothes = clothesRepository.findBySystemClothesId(systemClothesId).get();
+        Optional<Clothes> optionalClothes = clothesRepository.findBySystemClothesId(systemClothesId);
+        if (optionalClothes.isEmpty()) {
+            return null;
+        }
+        Clothes clothes = optionalClothes.get();
         Image aiImageByPrompt = imageService.createAiImageByPrompt(clothes, style, prompt);
         clothes.addImageFile(aiImageByPrompt);
         return aiImageByPrompt;
@@ -105,7 +115,11 @@ public class ClothesServiceImpl implements ClothesService {
 
     @Override
     public Image removeImageBackground(Long systemClothesId, String clothesName, String fileUrl) {
-        Clothes clothes = clothesRepository.findBySystemClothesId(systemClothesId).get();
+        Optional<Clothes> optionalClothes = clothesRepository.findBySystemClothesId(systemClothesId);
+        if (optionalClothes.isEmpty()) {
+            return null;
+        }
+        Clothes clothes = optionalClothes.get();
         Image removedImage = imageService.removeImageBackground(clothes, fileUrl);
         clothes.addImageFile(removedImage);
         return removedImage;
